smt-tomcat-junit-runner-parent/smt-tomcat8-junit-runner/src/main/java/shiver/me/timbers/junit/runner/tomcat/NullJarScanner.java
package shiver.me.timbers.junit.runner.tomcat;

import org.apache.tomcat.JarScanFilter;
import org.apache.tomcat.JarScanType;
import org.apache.tomcat.JarScanner;
import org.apache.tomcat.JarScannerCallback;

import javax.servlet.ServletContext;

/**
 * @author Karl Bennett
 */
public class NullJarScanner implements JarScanner {

    @Override
    public void scan(JarScanType scanType, ServletContext context, JarScannerCallback callback) {
    }

    @Override
    public JarScanFilter getJarScanFilter() {

        return new JarScanFilter() {
            @Override
            public boolean check(JarScanType jarScanType, String jarName) {
                return false;
            }
        };
    }

    @Override
    public void setJarScanFilter(JarScanFilter jarScanFilter) {
    }
}
